The ultimate driving machine
That is a bit of a claim! This version uses data augmentation techniques
- zoom, pan, brightness and flipping to augment the base data using a
custom image generator. In order to run this code you will need to
pip install imgaug which is the image augmentation library used.
See moodle for links to screencasts which explain this code and the
prvious commits which were not covered in class